moves decrosstalk into modules
