+++ tests/modules/decrosstalk/test_run_decrosstalk.py
@@ -1,6 +1,6 @@
 import os
-from ophys_etl.decrosstalk.ophys_plane import DecrosstalkingOphysPlane
-from ophys_etl.decrosstalk.decrosstalk import run_decrosstalk
+from ophys_etl.modules.decrosstalk.ophys_plane import DecrosstalkingOphysPlane
+from ophys_etl.modules.decrosstalk.decrosstalk import run_decrosstalk
 
 from .utils import create_data
